Misc changes to the client tests
